+++ setup.cfg
@@ -1,9 +1,9 @@
 [metadata]
 name = DMDO
-version = 1.3.0
+version = 2401
 author = Ahmed H. Bayoumy
 author_email = ahmed.bayoumy@mail.mcgill.ca
-description = "Python package for DMDO; an implementation of the NoHiMDO and ADMM."
+description = "Python package for DMDO; an implementation of the Nonhierarchical Analytical Target Cascading (NHATC.)"
 long_description = file: README.md
 long_description_content_type = text/markdown
 url = https://github.com/Ahmed-Bayoumy/DMDO
