+++ tests/test_files/Basic_MDO.yaml
@@ -1,14 +1,14 @@
 # The basic MDO problem
 variables:
 # vs:  [name, SP_index, Link, coupling type, lb, bl, ub, dim]
-  v0:  ["u",     1,       2,          s, 0., 1., 10.,  1]
-  v1:  ["v",     1,    None,          u, 0., 1., 10.,  1]
-  v2:  ["a",     1,       2,         ff, 0., 1., 10.,  1]
-  v3:  ["b",     1,       2,         fb, 0., 1., 10.,  1]
-  v4:  ["u",     2,       1,          s, 0., 1., 10.,  1]
-  v5:  ["w",     2,    None,          u, 0., 1., 10.,  1]
-  v6:  ["a",     2,       1,         fb, 0., 1., 10.,  1]
-  v7:  ["b",     2,       1,         ff, 0., 1., 10.,  1]
+  v0:  ["u",     1,       2,          s, 0., 1., 10.,  1, 'R']
+  v1:  ["v",     1,    None,          u, 0., 1., 10.,  1, 'R']
+  v2:  ["a",     1,       2,         ff, 0., 1., 10.,  1, 'R']
+  v3:  ["b",     1,       2,         fb, 0., 1., 10.,  1, 'R']
+  v4:  ["u",     2,       1,          s, 0., 1., 10.,  1, 'R']
+  v5:  ["w",     2,    None,          u, 0., 1., 10.,  1, 'R']
+  v6:  ["a",     2,       1,         fb, 0., 1., 10.,  1, 'R']
+  v7:  ["b",     2,       1,         ff, 0., 1., 10.,  1, 'R']
 
 DA:
   DA1:
