+++ tests/test_files/geometric_programming.yaml
@@ -1,23 +1,14 @@
 # The basic MDO problem
 variables:
-# vs:  [name, SP_index, Link, coupling type, lb,   bl,      ub, dim]
-  v0:  ["SFC",     1,       2,          fb, 0.,    1.,      4.,  1]
-  v1:  ["We",      1,       2,          fb, 0.,  100.,  30000.,  1]
-  v2:  ["LD",      1,       3,          fb, 0.,    .1,     10.,  1]
-  v3:  ["Ws",      1,       4,          fb, 0., 5000., 100000.,  1]
-  v4:  ["Wf",      1,       4,          fb, 0., 5000., 100000.,  1]
-  v5:  ["Wt",      1,       3,          ff, 0., 5000., 100000.,  1]
-
-  v6:  ["D",       2,       3,          fb, 0., 1000.,  70000.,  1]
-  v7:  ["T",       2,    None,           u, 0.,    .1,      1.,  1]
-  v8:  ["We",      2,       1,          ff, 0.,  100.,  30000.,  1]
-  v9:  ["SFC",     2,       1,          ff, 0.,    1.,      4.,  1]
-  v10: ["ESF",     2,       3,          ff, 0.,   0.5,    1.5.,  1]
-
-  v11: ["ESF",      2,       1,          s, 0., 1., 10.,  1]
-  v12: ["Wt",       2,    None,          u, 0., 1., 10.,  1]
-  v13: ["theta",    2,       1,         fb, 0., 1., 10.,  1]
-  v14: ["tc",       2,       1,         ff, 0., 1., 10.,  1]
+# vs:  [name, SP_index, Link, coupling type, lb, bl, ub, dim]
+  v0:  ["u",     1,       2,          s, 0., 1., 10.,  1, 'R']
+  v1:  ["v",     1,    None,          u, 0., 1., 10.,  1, 'R']
+  v2:  ["a",     1,       2,         ff, 0., 1., 10.,  1, 'R']
+  v3:  ["b",     1,       2,         fb, 0., 1., 10.,  1, 'R']
+  v4:  ["u",     2,       1,          s, 0., 1., 10.,  1, 'R']
+  v5:  ["w",     2,    None,          u, 0., 1., 10.,  1, 'R']
+  v6:  ["a",     2,       1,         fb, 0., 1., 10.,  1, 'R']
+  v7:  ["b",     2,       1,         ff, 0., 1., 10.,  1, 'R']
 
 DA:
   DA1:
